Move SubscriberInfo directory into SubscriberSection directory
