+++ src/components/admin/Settings/PrefixAlerts/Dialog/SubscriberSection/SubscriberInfo/index.tsx
@@ -2,9 +2,9 @@ import type { SubscriberAccordionProps } from 'src/components/admin/Settings/Pre
 
 import { Accordion, Stack, useTheme } from '@mui/material';
 
-import DeleteButton from 'src/components/admin/Settings/PrefixAlerts/Dialog/SubscriberInfo/DeleteButton';
-import Details from 'src/components/admin/Settings/PrefixAlerts/Dialog/SubscriberInfo/Details';
-import Summary from 'src/components/admin/Settings/PrefixAlerts/Dialog/SubscriberInfo/Summary';
+import DeleteButton from 'src/components/admin/Settings/PrefixAlerts/Dialog/SubscriberSection/SubscriberInfo/DeleteButton';
+import Details from 'src/components/admin/Settings/PrefixAlerts/Dialog/SubscriberSection/SubscriberInfo/Details';
+import Summary from 'src/components/admin/Settings/PrefixAlerts/Dialog/SubscriberSection/SubscriberInfo/Summary';
 import useAlertSubscriptionsStore from 'src/components/admin/Settings/PrefixAlerts/useAlertSubscriptionsStore';
 import { useEvaluateSubscriptionIneligibility } from 'src/components/admin/Settings/PrefixAlerts/useEvaluateSubscriptionIneligibility';
 import { defaultOutline, defaultOutlineColor_hovered } from 'src/context/Theme';
